+++ opencascade-hs/cpp/hs_types.h
@@ -71,6 +71,8 @@ typedef void Bnd_Box;
 typedef void Bnd_OBB;
 typedef void Message_ProgressRange;
 typedef void RWGltf_CafWriter;
+typedef void RWGltf_CafReader;
+typedef void RWMesh_CafReader;
 typedef void TColStd_IndexedDataMapOfStringString;
 typedef void TDF_Label;
 typedef void TDocStd_Document;