+++ opencascade-hs/src/OpenCascade/RWMesh/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.RWMesh.Types 
+( CafReader
+) where
+
+data CafReader